back-end/src/test/java/com/eventmanager/pachanga/factory/RelatorioEstoqueTOFactoryTest.java
package com.eventmanager.pachanga.factory;

import static org.junit.Assert.assertEquals;

import java.time.LocalDateTime;
import java.util.LinkedHashMap;

import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.security.oauth2.provider.token.AuthorizationServerTokenServices;
import org.springframework.test.context.junit4.SpringRunner;

import com.eventmanager.pachanga.dtos.RelatorioEstoqueTO;
import com.eventmanager.pachanga.securingweb.JwtAuthenticationEntryPoint;
import com.eventmanager.pachanga.securingweb.JwtTokenUtil;
import com.eventmanager.pachanga.securingweb.JwtUserDetailsService;

@RunWith(SpringRunner.class)
@WebMvcTest(value=RelatorioEstoqueTOFactory.class)
class RelatorioEstoqueTOFactoryTest {
	
	@MockBean
	private AuthorizationServerTokenServices defaultAuthorizationServerTokenServices;
	
	@MockBean
	private JwtUserDetailsService defaultJwtUserDetailsService;
	
	@MockBean
	private JwtTokenUtil defaultJwtTokenUtil;
	
	@Autowired
	private RelatorioEstoqueTOFactory relatorioEstoqueTOFactory;
	
	@MockBean
	private JwtAuthenticationEntryPoint defaultJwtAuthenticationEntryPoint;
	
	@Test
	void getRelatorioEstoqueTest() {
		
		RelatorioEstoqueTO relatorio = relatorioEstoqueTOFactory.getRelatorioEstoque("teste", "teste123", new LinkedHashMap<LocalDateTime, Integer>());
		
		assertEquals(true, relatorio.getQuantidadeHora() != null);
		assertEquals(true, "teste".equals(relatorio.getNomeEstoque()));
		
	}

}
